feat: fix AST signature for module setup
