+++ pykokkos/core/visitors/workunit_visitor.py
@@ -4,38 +4,15 @@ from typing import Dict, List, Optional, Set, Tuple, Union
 
 from pykokkos.core import cppast
 from pykokkos.core.keywords import Keywords
+from pykokkos.core.translators.reducer_util import (
+    NON_SCALAR_REDUCERS,
+    get_cpp_type_name,
+    get_reducer_value_type,
+)
 from pykokkos.interface import TeamMember
 
 from . import visitors_util
 from .pykokkos_visitor import PyKokkosVisitor
-
-
-VALUE_LOC_REDUCERS = {"MaxLoc", "MinLoc"}
-MINMAX_REDUCERS = {"MinMax"}
-MINMAX_LOC_REDUCERS = {"MinMaxLoc"}
-NON_SCALAR_REDUCERS = VALUE_LOC_REDUCERS | MINMAX_REDUCERS | MINMAX_LOC_REDUCERS
-
-
-def get_cpp_type_name(decltype: cppast.Type) -> str:
-    if isinstance(decltype, cppast.PrimitiveType):
-        typename = decltype.typename
-        return typename.value if hasattr(typename, "value") else typename
-
-    if isinstance(decltype, cppast.ClassType):
-        return decltype.typename
-
-    raise TypeError(f"Unsupported reducer value type: {decltype}")
-
-
-def get_reducer_value_type(reducer: str, value_type: str) -> str:
-    if reducer in VALUE_LOC_REDUCERS:
-        return f"Kokkos::{reducer}<{value_type},int>::value_type"
-    if reducer in MINMAX_REDUCERS:
-        return f"Kokkos::{reducer}<{value_type}>::value_type"
-    if reducer in MINMAX_LOC_REDUCERS:
-        return f"Kokkos::{reducer}<{value_type},int>::value_type"
-
-    raise ValueError(f"Reducer {reducer} does not use a non-scalar value_type")
 
 
 class WorkunitVisitor(PyKokkosVisitor):
